requirements.txt
asgiref==3.8.1
attrs==25.3.0
boto3==1.37.20
botocore==1.37.20
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==2.3.0
Django==5.2
django-autoslug==1.9.9
django-ckeditor==6.7.2
django-debug-toolbar==5.2.0
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.2
django-ratelimit==4.1.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
psycopg==3.2.3
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
s3transfer==0.11.5
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
